Modify model manager to load / unload model based on dependency. Enable ... (#460)
* Enable and fixed ensemble validation test

Fix ensemble in L0_model_config

* Modify ensemble utils to accommodate to dependency graph change

* Modify model repository manager to load model based on dependency

* Remove skip in reshape test as the reshape in ensemble model is fixed

+++ qa/L0_model_config/test.sh
@@ -57,45 +57,45 @@ for modelpath in \
        $modelpath/.
 done
 
-# The tests in autofill_noplatform/ensemble_norun cannot be run until
-# DLIS-519 is resolved. Once that issue is resolve the tests should be
-# moved into autofill_noplatform/ensemble and the following
-# initialization should be uncommented.
-rm -fr autofill_noplatform/ensemble_norun
-#for modelpath in \
-#    autofill_noplatform/ensemble/invalid_input_map/invalid_input_map/1 \
-#        autofill_noplatform/ensemble/invalid_input_map/fp32_dim1_batch4/1 \
-#        autofill_noplatform/ensemble/invalid_input_map/fp32_dim1_batch4_input4/1 \
-#        autofill_noplatform/ensemble/invalid_input_map/fp32_dim1_batch4_output3/1 \
-#        autofill_noplatform/ensemble/invalid_output_map/invalid_output_map/1 \
-#        autofill_noplatform/ensemble/invalid_output_map/fp32_dim1_batch4/1 \
-#        autofill_noplatform/ensemble/invalid_output_map/fp32_dim1_batch4_input4/1 \
-#        autofill_noplatform/ensemble/invalid_output_map/fp32_dim1_batch4_output3/1 \
-#        autofill_noplatform/ensemble/invalid_batch_size/invalid_batch_size/1 \
-#        autofill_noplatform/ensemble/invalid_batch_size/invalid_batch_size/1 \
-#        autofill_noplatform/ensemble/invalid_batch_size/fp32_dim1_batch2/1 \
-#        autofill_noplatform/ensemble/invalid_batch_size/fp32_dim1_batch4/1 \
-#        autofill_noplatform/ensemble/inconsistent_shape/inconsistent_shape/1 \
-#        autofill_noplatform/ensemble/inconsistent_shape/fp32_dim1_batch4/1 \
-#        autofill_noplatform/ensemble/inconsistent_shape/fp32_dim3_batch4/1 \
-#        autofill_noplatform/ensemble/inconsistent_data_type/inconsistent_data_type/1 \
-#        autofill_noplatform/ensemble/inconsistent_data_type/fp32_dim1_batch2/1 \
-#        autofill_noplatform/ensemble/inconsistent_data_type/int32_dim1_batch4/1 \
-#        autofill_noplatform/ensemble/embedded_ensemble/ensemble/1 \
-#        autofill_noplatform/ensemble/embedded_ensemble/fp32_dim1_batch4/1 \
-#        autofill_noplatform/ensemble/embedded_ensemble/inner_ensemble/1 \
-#        autofill_noplatform/ensemble/non_existing_model/non_existing_model/1 \
-#        autofill_noplatform/ensemble/non_existing_model/fp32_dim1_batch4/1 \
-#        autofill_noplatform/ensemble/non_existing_model/fp32_dim1_batch4_output3/1 \
-#        autofill_noplatform/ensemble/self_circular_dependency/self_circular_dependency/1 \
-#        autofill_noplatform/ensemble/self_circular_dependency/fp32_dim1_batch4/1 \
-#        autofill_noplatform/ensemble/self_circular_dependency/fp32_dim1_batch4_input4/1 \
-#        autofill_noplatform/ensemble/self_circular_dependency/fp32_dim1_batch4_output3/1 \
-#        autofill_noplatform/ensemble/circular_dependency/circular_dependency/1 \
-#        autofill_noplatform/ensemble/circular_dependency/circular_dependency_2/1 ; do
-#    mkdir -p $modelpath
-#    cp ./libidentity.so $modelpath/libcustom.so
-#done
+for modelpath in \
+   autofill_noplatform/ensemble/invalid_input_map/invalid_input_map/1 \
+       autofill_noplatform/ensemble/invalid_input_map/fp32_dim1_batch4/1 \
+       autofill_noplatform/ensemble/invalid_input_map/fp32_dim1_batch4_input4/1 \
+       autofill_noplatform/ensemble/invalid_input_map/fp32_dim1_batch4_output3/1 \
+       autofill_noplatform/ensemble/invalid_output_map/invalid_output_map/1 \
+       autofill_noplatform/ensemble/invalid_output_map/fp32_dim1_batch4/1 \
+       autofill_noplatform/ensemble/invalid_output_map/fp32_dim1_batch4_input4/1 \
+       autofill_noplatform/ensemble/invalid_output_map/fp32_dim1_batch4_output3/1 \
+       autofill_noplatform/ensemble/invalid_batch_size/invalid_batch_size/1 \
+       autofill_noplatform/ensemble/invalid_batch_size/invalid_batch_size/1 \
+       autofill_noplatform/ensemble/invalid_batch_size/fp32_dim1_batch2/1 \
+       autofill_noplatform/ensemble/invalid_batch_size/fp32_dim1_batch4/1 \
+       autofill_noplatform/ensemble/inconsistent_shape/inconsistent_shape/1 \
+       autofill_noplatform/ensemble/inconsistent_shape/fp32_dim1_batch4/1 \
+       autofill_noplatform/ensemble/inconsistent_shape/fp32_dim3_batch4/1 \
+       autofill_noplatform/ensemble/inconsistent_data_type/inconsistent_data_type/1 \
+       autofill_noplatform/ensemble/inconsistent_data_type/fp32_dim1_batch2/1 \
+       autofill_noplatform/ensemble/inconsistent_data_type/int32_dim1_batch4/1 \
+       autofill_noplatform/ensemble/non_existing_model/non_existing_model/1 \
+       autofill_noplatform/ensemble/non_existing_model/fp32_dim1_batch4/1 \
+       autofill_noplatform/ensemble/non_existing_model/fp32_dim1_batch4_output3/1 \
+       autofill_noplatform/ensemble/self_circular_dependency/self_circular_dependency/1 \
+       autofill_noplatform/ensemble/self_circular_dependency/fp32_dim1_batch4/1 \
+       autofill_noplatform/ensemble/self_circular_dependency/fp32_dim1_batch4_input4/1 \
+       autofill_noplatform/ensemble/self_circular_dependency/fp32_dim1_batch4_output3/1 \
+       autofill_noplatform/ensemble/unmapped_input/unmapped_input/1 \
+       autofill_noplatform/ensemble/unmapped_input/fp32_dim1_batch4/1 \
+       autofill_noplatform/ensemble/unmapped_input/fp32_dim1_batch4_input4/1 \
+       autofill_noplatform/ensemble/circular_dependency/circular_dependency/1 \
+       autofill_noplatform/ensemble/circular_dependency/circular_dependency_2/1 \
+       autofill_noplatform_success/ensemble/embedded_ensemble/embedded_ensemble/1 \
+       autofill_noplatform_success/ensemble/embedded_ensemble/fp32_dim1_batch4/1 \
+       autofill_noplatform_success/ensemble/embedded_ensemble/inner_ensemble/1 \
+       autofill_noplatform_success/ensemble/unmapped_output/unmapped_output/1 \
+       autofill_noplatform_success/ensemble/unmapped_output/fp32_dim1_batch4_output3/1 ; do
+   mkdir -p $modelpath
+   cp ./libidentity.so $modelpath/libcustom.so
+done
 
 
 rm -f $SERVER_LOG_BASE* $CLIENT_LOG
@@ -167,11 +167,11 @@ for TARGET_DIR in `ls -d autofill_noplatform/*/*`; do
     rm -fr models && mkdir models
     if [ -f ${TARGET_DIR}/config.pbtxt ]; then
         cp -r ${TARGET_DIR} models/.
+        EXPECTEDS=models/$TARGET/expected*
     else
         cp -r ${TARGET_DIR}/* models/.
+        EXPECTEDS=models/expected*
     fi
-
-    EXPECTEDS=models/$TARGET/expected*
 
     echo -e "Test ${TARGET_DIR}" >> $CLIENT_LOG
 
@@ -212,8 +212,15 @@ for TARGET_DIR in `ls -d autofill_noplatform_success/*/*`; do
     SERVER_ARGS="--model-store=`pwd`/models --strict-model-config=false"
     SERVER_LOG=$SERVER_LOG_BASE.${TARGET_DIR_DOT}.log
 
+    # If there is a config.pbtxt at the top-level of the test then
+    # assume that the directory is a single model. Otherwise assume
+    # that the directory is an entire model repository.
     rm -fr models && mkdir models
-    cp -r ${TARGET_DIR} models/.
+    if [ -f ${TARGET_DIR}/config.pbtxt ] || [ "$TARGET" = "no_config" ]; then
+        cp -r ${TARGET_DIR} models/.
+    else
+        cp -r ${TARGET_DIR}/* models/.
+    fi
 
     echo -e "Test $TARGET_DIR" >> $CLIENT_LOG
